refactor: moved scene graph ui code out of common
