+++ common/Cargo.toml
@@ -45,5 +45,4 @@ petgraph = { version = "0.8.1", default-features = false, features = [
 proc-macros = { version = "0.1.0", path = "../proc-macros" }
 fxhash = "0.2.1"
 nalgebra = { version = "0.34.1", features = ["serde", "serde-serialize"] }
-egui_ltreeview = "0.6.0"
 fastrand = "2.3.0"
