+++ editor/Cargo.toml
@@ -22,3 +22,9 @@ glium = "0.36"
 egui_glium = "0.33.2"
 serde_json = "1.0.116"
 nalgebra = "0.34.1"
+egui_ltreeview = "0.6.0"
+petgraph = { version = "0.8.1", default-features = false, features = [
+    "serde-1",
+    "stable_graph",
+] }
+itertools = "0.14.0"
