+++ editor/src/main.rs
@@ -4,6 +4,7 @@ use common::run;
 use editor::Editor;
 
 mod editor;
+mod ui;
 
 fn main() {
     let attributes = WindowAttributes::default().with_name("shooter-game-editor", "");
