+++ editor/src/ui/mod.rs
@@ -1,0 +1,7 @@
+mod ui_graph;
+
+use egui_glium::egui_winit::egui::Ui;
+
+pub trait Show {
+    fn show(&mut self, ui: &mut Ui);
+}
